Consolidate events tests
