+++ tests/integration/test_workflow_events.py
@@ -15,12 +15,13 @@
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 import asyncio
+from types import MethodType
 
 from async_timeout import timeout as async_timeout
 import pytest
-from types import MethodType
 
 from cylc.flow.scheduler import SchedulerError
+from cylc.flow.workflow_events import WorkflowEventHandler
 
 
 EVENTS = (
@@ -55,13 +56,6 @@ async def test_scheduler(flow, scheduler, capcall):
             'scheduling': {
                 'graph': {
                     'R1': 'a'
-                }
-            },
-            'runtime': {
-                'a': {
-                    'simulation': {
-                        'default run length': 'PT0S',
-                    }
                 }
             },
         })
@@ -234,3 +228,104 @@ async def test_shutdown_handler_timeout_kill(
         "[('workflow-event-handler-00', 'shutdown') err] killed on "
         "timeout (PT1S)"
     ) in caplog.text
+
+
+TEMPLATES = [
+    # perfectly valid
+    pytest.param('%(workflow)s', id='good'),
+    # no template variable of that name
+    pytest.param('%(no_such_variable)s', id='bad'),
+    # missing the 's'
+    pytest.param('%(broken_syntax)', id='ugly'),
+]
+
+
+@pytest.mark.parametrize('template', TEMPLATES)
+async def test_mail_footer_template(
+    mod_one,  # use the same scheduler for each test
+    start,
+    mock_glbl_cfg,
+    log_filter,
+    capcall,
+    template,
+):
+    """It should handle templating issues with the mail footer."""
+    # prevent emails from being sent
+    mail_calls = capcall(
+        'cylc.flow.workflow_events.WorkflowEventHandler._send_mail'
+    )
+
+    # configure Cylc to send an email on startup with the configured footer
+    mock_glbl_cfg(
+        'cylc.flow.workflow_events.glbl_cfg',
+        f'''
+            [scheduler]
+                [[mail]]
+                    footer = 'footer={template}'
+                [[events]]
+                    mail events = startup
+        ''',
+    )
+
+    # start the workflow and get it to send an email
+    async with start(mod_one) as one_log:
+        one_log.clear()  # clear previous log messages
+        mod_one.workflow_event_handler.handle(
+            mod_one,
+            WorkflowEventHandler.EVENT_STARTUP,
+            'event message'
+        )
+
+    # warnings should appear only when the template is invalid
+    should_log = 'workflow' not in template
+
+    # check that template issues are handled correctly
+    assert bool(log_filter(
+        contains='Ignoring bad mail footer template',
+    )) == should_log
+    assert bool(log_filter(
+        contains=template,
+    )) == should_log
+
+    # check that the mail is sent even if there are issues with the footer
+    assert len(mail_calls) == 1
+
+
+@pytest.mark.parametrize('template', TEMPLATES)
+async def test_custom_event_handler_template(
+    mod_one,  # use the same scheduler for each test
+    start,
+    mock_glbl_cfg,
+    log_filter,
+    template,
+):
+    """It should handle templating issues with custom event handlers."""
+    # configure Cylc to send an email on startup with the configured footer
+    mock_glbl_cfg(
+        'cylc.flow.workflow_events.glbl_cfg',
+        f'''
+            [scheduler]
+                [[events]]
+                    startup handlers = echo "{template}"
+        '''
+    )
+
+    # start the workflow and get it to send an email
+    async with start(mod_one) as one_log:
+        one_log.clear()  # clear previous log messages
+        mod_one.workflow_event_handler.handle(
+            mod_one,
+            WorkflowEventHandler.EVENT_STARTUP,
+            'event message'
+        )
+
+    # warnings should appear only when the template is invalid
+    should_log = 'workflow' not in template
+
+    # check that template issues are handled correctly
+    assert bool(log_filter(
+        contains='bad template',
+    )) == should_log
+    assert bool(log_filter(
+        contains=template,
+    )) == should_log
